canteen-management-system/backend/requirements.txt
django>=5.2
djangorestframework
mysqlclient
django-cors-headers
django-filter
